+++ src/Carousel.js
@@ -19,18 +19,93 @@ class CarouselIthreads extends HTMLElement {
   }
 
   _getDefaultTemplate() {
+    const styles = this._getDefaultStyles();
+    return `
+${styles}
+<div part="container-slot">
+  <slot></slot>
+</div>
+<div part="container-grid">
+  <button id="btnPrev" part="button previous">Previous</button>
+  <button id="btnNext" part="button next">Next</button>
+  <ul id="pagination" part="container-pagination"></ul>
+</div>
+`;
+  }
+
+  _getDefaultStyles() {
     return `
 <style>
+
 :host {
   display: block;
 }
+
+/* Organize the nav buttons and pagination in a grid. */
+[part="container-grid"] {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  grid-template-rows: 1fr;
+  gap: 0px .5em;
+  grid-template-areas:
+    "prev next";
+}
+
+/* If we're showing pagination, insert them into the grid. */
+:host([pagination]) [part="container-grid"] {
+  grid-template-columns: 1fr max-content 1fr;
+  grid-template-areas:
+    "prev pagination next";
+}
+
+[part="button previous"] {
+  justify-self: end;
+  grid-area: prev;
+}
+
+[part="button next"] { 
+  justify-self: start;
+  grid-area: next; 
+}
+
+[part="container-pagination"] {
+  justify-self: center;
+  align-self: center;
+  grid-area: pagination;
+}
+
+[part~="button"] {
+  width: 8em;
+}
+
+[part="container-pagination"] {
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+}
+
+[part~="pagination-item"] {
+  display: inline-block;
+}
+
+[part~="pagination-button"] {
+  border-radius: .75em;
+  border: 1px solid #ccc;
+  margin: 0 .1em;
+}
+
+[part="pagination-button"] {
+  background: ButtonFace;
+  color: inherit;
+}
+
+[part="pagination-button active"] {
+  background: black;
+  color: white;
+}
+
 </style>
-<div id="container">
-  <slot></slot>
-</div>
-<button id="btnPrev">Previous</button>
-<button id="btnNext">Next</button>
-`;
+    `;
   }
 
   connectedCallback() {
@@ -101,6 +176,42 @@ class CarouselIthreads extends HTMLElement {
     }
   }
 
+  _addPagination() {
+    const el = this.shadowRoot.getElementById("pagination");
+    if (!el) return;
+
+    for (let i = 0; i < this.state.total; i++) {
+      const li = document.createElement("li");
+      li.setAttribute("part", "pagination-item");
+
+      const button = document.createElement("button");
+      button.setAttribute("data-item", i);
+      button.setAttribute("aria-label", "Go to item " + (i + 1) + " of " + this.state.total);
+      button.setAttribute("part", "pagination-button");
+      button.textContent = "" + (i + 1);
+      button.addEventListener("click", this._goto.bind(this));
+
+      li.appendChild(button);
+      el.appendChild(li);
+    }
+  }
+
+  _updatePagination(newIndex) {
+    const elPagination = this.shadowRoot.getElementById("pagination");
+    if (elPagination) {
+      elPagination.querySelectorAll("li").forEach((elItem, index) => {
+        const elButton = elItem.querySelector("button");
+        if (index === newIndex) {
+          elItem.setAttribute("part", "pagination-item active");
+          elButton.setAttribute("part", "pagination-button active");
+        } else {
+          elItem.setAttribute("part", "pagination-item");
+          elButton.setAttribute("part", "pagination-button");
+        }
+      });
+    }
+  }
+
   onMutation(mutations) {
     const added = [];
 
@@ -119,6 +230,10 @@ class CarouselIthreads extends HTMLElement {
       return el.nodeType === Node.ELEMENT_NODE;
     }).length;
 
+    if (this.hasAttribute("pagination")) {
+      this._addPagination();
+    }
+
     this.activate(this.state.current);
   }
 
@@ -134,6 +249,11 @@ class CarouselIthreads extends HTMLElement {
    */
   next() {
     return this.activate(this.state.current + 1);
+  }
+
+  _goto(event) {
+    const index = parseInt(event.target.getAttribute("data-item"), 10);
+    this.activate(index);
   }
 
   /**
@@ -156,6 +276,10 @@ class CarouselIthreads extends HTMLElement {
         el.setAttribute("hidden", "hidden");
       }
     });
+
+    // If we're displaying pagination buttons, update attributes to
+    // indicate which one is currently active.
+    this._updatePagination(newIndex);
 
     return this.state.current;
   }